Refactor JobInfoList into own component
