+++ src/socket.js
@@ -1,0 +1,5 @@
+import { io } from 'socket.io-client';
+
+const socket = io(`${import.meta.env.VITE_CEA_URL}`);
+
+export default socket;
